move example script out of package
